# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para upload, download e conversão de arquivos mp3 para ogg",
        "title": "Simple File Redirect API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/manager/v1/convert": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Realiza upload e conversão de um arquivo MP3 para OGG",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conversão"
                ],
                "summary": "Conversão de arquivo MP3 para OGG",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Arquivo MP3 para conversão",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Extensão de destino (ex: ogg)",
                        "name": "convert",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo convertido com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação ou tipo de conversão não suportado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao converter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/manager/v1/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Realiza o download de um arquivo salvo, baseado no path informado",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Arquivos"
                ],
                "summary": "Download de arquivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Caminho completo do arquivo salvo",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo enviado",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Parâmetro ausente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/manager/v1/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Recebe um arquivo via multipart/form e salva no diretório de arquivos",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Arquivos"
                ],
                "summary": "Upload de arquivo",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Arquivo para upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload realizado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao salvar",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Forneça o token no formato: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}